Change dict key from random val to index
- The key for dict was initially a variable
which was initialized to zero; this value was
incremented by a counter. This allowed for
dupes to be added to the watchlist
- Duplicates are prevented by changing the
key to the film index; since this value is
unique, if a user choses a dupe index, it
will just overwrite with the same value,
which effectively causes no change.