Add convert_dict_2_tuples, create_unique_film_list
- Convert the list of films to watch from a list
of lists of Film objects to a list of lists, with
each element in the list being a tuple contaning
the film name, and the Screening object; this is
done using the 'convert_dict_2_tuples' function.
- Create a function 'create_unique_film_list'
which creates the cartesian product using the
film tuples. This cartesian product creates a
list of lists of all possible combinations a user
can watch the films if they are to see only one
screening per film.
- Add type hinting to all functions.